corrections
adding the pointer property in the selector box
operational correction in the selector box
add testings
flexbox operation correction in the rendered list
margin and padding corrections
removal of css properties that are repeated on different layers.
min-wide correction of "li" items to achieve better visual behavior.
correction of the key that caused a failure because the list item was wrapped in a fragment